+++ lib/Controller/GiftController.dart
@@ -1,6 +1,7 @@
 import 'package:hedieaty/Firebase/FirebaseGiftService.dart';
 import 'package:hedieaty/Model/Gift.dart';
 import 'package:hedieaty/SQL_Local/LocalDB.dart';
+import 'package:hedieaty/main.dart';
 
 class GiftController {
   static final LocalDB _localDB = LocalDB();
@@ -8,8 +9,11 @@ class GiftController {
 
   static Future<bool> addGift(Gift gift) async {
     try {
+      gift.isPublished = autoSync == 1 ? 1 : 0;
       await _localDB.insertGift(gift);
-      await _firebaseGiftService.createGift(gift);
+      if (autoSync == 1) {
+        await _firebaseGiftService.createGift(gift);
+      }
       return true;
     } catch (e) {
       print("Error: $e");
@@ -19,16 +23,23 @@ class GiftController {
 
   static Future<bool> updateGift(Gift gift) async {
     try {
+      gift.isPublished = autoSync == 1 ? 1 : 0;
       await _localDB.updateGift(gift);
-      await _firebaseGiftService.updateGift(gift);
+      if (autoSync == 1) {
+        await _firebaseGiftService.updateGift(gift);
+      }
       return true;
     } catch (e) {
-      print("Error: \$e");
+      print("Error: $e");
       return false;
     }
   }
 
   static Future<bool> deleteGift(String id) async {
+    if (autoSync == 0) {
+      print("Delete not allowed when auto-sync is off.");
+      return false;
+    }
     try {
       await _localDB.deleteGift(id);
       await _firebaseGiftService.deleteGift(id);
@@ -40,6 +51,10 @@ class GiftController {
   }
 
   static Future<bool> deleteAllGiftsByEventId(String id) async {
+    if (autoSync == 0) {
+      print("Delete not allowed when auto-sync is off.");
+      return false;
+    }
     try {
       await _localDB.deleteGiftsByEventId(id);
       await _firebaseGiftService.deleteGiftsByEventId(id);
